Safari-611.1.10/LayoutTests/fast/xsl/xslt-contentType-expected.txt

PASS doc.contentType is "text/html"
PASS doc.contentType is "application/xml"
PASS doc.contentType is "application/xhtml+xml"
PASS doc.contentType is "application/xml"
PASS doc.contentType is "application/xml"
PASS successfullyParsed is true

TEST COMPLETE
